+++ docker-compose/dev/apps/config/sampler.json
@@ -1,7 +1,7 @@
 {
   "postgres_uri": "postgresql://admin:admin@postgresql:5432/pocket-ml-testbench",
   "mongodb_uri": "mongodb://mongodb:27017/pocket-ml-testbench",
-  "log_level": "DEBUG",
+  "log_level": "INFO",
   "temporal": {
     "host": "temporal",
     "port": 7233,
