+++ docker-compose/dev/apps/docker-compose.yaml
@@ -30,7 +30,7 @@ services:
       EVALUATE_VERBOSITY: "error"
       DATASETS_VERBOSITY: "error"
     deploy:
-      replicas: 3
+      replicas: 5
     extra_hosts:
       host.docker.internal: host-gateway
     volumes:
